REALLY fix patch

+++ lambdas/src/routes/patch_dessert.py
@@ -90,24 +90,27 @@ def patch_dessert(request: Request, body: PatchDessertRequest, dessert_id: str):
         dessert_images_bucket = os.environ.get(
             "DESSERT_IMAGES_BUCKET_NAME", "pc-dessert-images-bucket-dev"
         )
-        if "images" in get_dessert_response["Item"]:
-            for image in get_dessert_response["Item"]["images"]:
-                s3_client.delete_object(
-                    Bucket=dessert_images_bucket,
-                    Key=f"{dessert_id}/{image['image_id']}",
-                )
-            logger.info(f"Deleted images for dessert: {dessert_id}")
 
         updated_images = []
         for image in updated_dessert.get("images"):
-            image_id = str(uuid.uuid4())
-            object_url = f"https://{dessert_images_bucket}.s3.amazonaws.com/{dessert_id}/{image_id}"
-            image["image_id"] = image_id
-            image["url"] = object_url
-            image["upload_url"] = generate_upload_url(
-                dessert_id, image, dessert_images_bucket
-            )
+            if "image_id" not in image:
+                image_id = str(uuid.uuid4())
+                object_url = f"https://{dessert_images_bucket}.s3.amazonaws.com/{dessert_id}/{image_id}"
+                image["image_id"] = image_id
+                image["url"] = object_url
+                image["upload_url"] = generate_upload_url(
+                    dessert_id, image, dessert_images_bucket
+                )
             updated_images.append(image)
+
+        if "images" in get_dessert_response["Item"]:
+            for image in get_dessert_response["Item"]["images"]:
+                if image["image_id"] not in [img["image_id"] for img in updated_images]:
+                    s3_client.delete_object(
+                        Bucket=dessert_images_bucket,
+                        Key=f"{dessert_id}/{image['image_id']}",
+                    )
+
         update_expression_data["images"] = updated_images
 
     if "prices" in updated_dessert:
